fix: 대화 기록 불러오기 함수 분리 및 타입 오류 해결
